MH.BLTest/CustomerRepositoryTest.cs
using System;
using MH.BL;
using Microsoft.VisualStudio.TestTools.UnitTesting;

namespace MH.BLTest
{
    [TestClass]
    public class CustomerRepositoryTest
    {
        [TestMethod]
        public void TestMethod1()
        {
            //Arrange
            var customerRepository = new CustomerRepository();
            var expected = new Customer(1)
            {
                EmailAddress = "Test@gmail.com",
                FirstName = "Nguyen",
                LastName = "Manh Hoang"
            };

            //Act
            var actual = customerRepository.Retrieve(1);

            //Assert
            Assert.AreEqual(expected.EmailAddress, actual.EmailAddress);
            Assert.AreEqual(expected.FirstName, actual.FirstName);
            Assert.AreEqual(expected.LastName, actual.LastName);
        }
    }
}
